Fix handling of aulow, hideclient_to and appendchaninfo settings.
These settings were loaded from [monitor] section in oscam.conf
but they are global in nature, used in monitor, webif, oscam logging
and other places.

Also when the new config parser was introduced, in case [monitor]
was disabled these values didn't get their default values which
causes strange behaviour of the webif that depended on them.

This commit fixes the issues by making the variables not depend
on [monitor] only. They are loaded from [monitor] and [webif] and
saved in both places (the last one wins). That way, old configs that
have set the values in [monitor] would continue to work.

This commit fixes tickets #2725 and #2739.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@7409 4b0bc96b-bc66-0410-9d44-ebda105a78c1