Add "ecmnotfoundlimit" reader setting.
ecmnotfoundlimit setting is compared to the number of ECM NOKs
returned by the reader. Once ECM NOKs are equal to ecmnotfoundlimit,
the reader is restarted.

Sometimes cards start answering with "not found" for some SIDs and
only restart of the reader causes them to operate normally. This
setting helps in that case.

The problem was observed with NDS SkyUK card.

That setting defaults to 0 (no limit).

git-svn-id: https://svn.streamboard.tv/oscam/trunk@8313 4b0bc96b-bc66-0410-9d44-ebda105a78c1