config: Remove some logging.
If config files do not exist we tell the user, then a user that
doesn't need these files just creates empty files and we start
to complain that they are not "in the right format". Its easier and
cleaner to just report when we have something to report otherwise
we should STFU.

Thanks to Gorgone for reminding me to remove these annoying logs.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@8378 4b0bc96b-bc66-0410-9d44-ebda105a78c1