[osemu + biss] Added support for BISS2 mode CA emm keys EMM keys for BISS2 mode CA are essentially RSA private keys.
They will be read from external PEM files and stored internally in OSCam-Emu. They will not be saved in the SoftCam.Key and will not be visible from insdie OSCam-Emu (there is no need for that).

For every RSA private key there is a unique identifier named "entitlement key id" (or ekid for short) which is derived form each key and is also present in the EMM. The ekid of each RSA key will be visible to the user via the entitlement list of the Emu reader.

In order to load PEM files containing RSA private keys to OSCam-emu, you need to name them in the "biss2_private_XX.pem" format where "XX" is a two digit number (from 00 to 16 at the moment). PEM files must be located to the same directory where the SoftCam.Key lies.